fix: change dialog to bottomsheet on mobile
