+++ lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:device_preview/device_preview.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -13,15 +14,18 @@ void main() async {
   /// Localization
   await EasyLocalization.ensureInitialized();
 
-  runApp(ProviderScope(
-      observers: [AppObserver()],
-      child: EasyLocalization(
-        supportedLocales: const [Locale('en')],
-        path: 'assets/translations',
-        fallbackLocale: const Locale('en'),
-        useOnlyLangCode: true,
-        child: MainApp(),
-      )));
+  runApp(DevicePreview(
+    enabled: true,
+    builder: (context) => ProviderScope(
+        observers: [AppObserver()],
+        child: EasyLocalization(
+          supportedLocales: const [Locale('en')],
+          path: 'assets/translations',
+          fallbackLocale: const Locale('en'),
+          useOnlyLangCode: true,
+          child: MainApp(),
+        )),
+  ));
 }
 
 class MainApp extends ConsumerWidget {
