+++ pubspec.yaml
@@ -42,6 +42,7 @@ dependencies:
   appcheck: ^1.5.2
   image_picker: ^1.1.2
   reown_appkit: ^1.3.0
+  device_preview: ^1.2.0
 
 dev_dependencies:
   flutter_test:
